chromium: update bbappend to 54.0.2810.2
The 52.0.2743.76 does not build with rocko, so update the recipe to
54.0.2810.2 which works with 678706cb43fb8576ff7b88bfb612993b8542724d
on meta-browser although 64.0.3282.167 builds fine with
a413b4f6608eac46b72a3a79dcada0767fe4cf58 on meta-browser.

We will insert a bbappend for 64.0.3282.167 in a separate commit.

While at it split the autostart service into a seperate recipe
and fix a race condition between chromium-autostart.service and
xserver-nodm.service.

Signed-off-by: Julia Kutter <jkutter@christ-es.de>
Acked-by: Peter Fink <pfink@christ-es.de>
Signed-off-by: Peter Fink <pfink@christ-es.de>