mesa: bump version of mesa to 19.1.7 locally
A few files to support a newer version of mesa are copied from
meta-openembedded master branch, it includes mesa.inc, the needed
patches for building and mesa_19.1.6.bb recipe, the version of
mesa software is forward ported to 19.1.7 at the same time.

The new version of mesa contains a number of essential fixes to
support etnaviv backend.

Signed-off-by: Vladimir Zapolskiy <vz@ilbers.de>