systemd-conf: select and install a locale at build time
Commonly systemd-firstboot service is supposed to set a locale,
however, because systemd-conf package brings /etc/machine-id file
to the root filesystem, the systemd-firstboot service is NOT run
on the first boot and machine locale setting remains unconfigured.

There are various options to configure a wanted locale in runtime,
for instance by running "localectl", but at build time the best
way is to install /etc/machine-id and /etc/locale.conf together.

The change fixes an issue with running tmux-2.6, which strictly
requires LANG variable to point to some installed UTF-8 locale.

Signed-off-by: Vladimir Zapolskiy <vz@ilbers.de>
Signed-off-by: Steffen Kothe <skothe@christ-es.de>