qt5: webengine: add fix for rotation bug and fix settings file for eglfs
The introduced patch is only for use with meta-qt5 layer from:
https://code.qt.io/yocto/meta-qt5.git

The introduced patch, allows rotation without buggy <select> elements
in the html code when displayed with qtwebengine.

The mainline qt5 layer introduced the bug, that a rotation causes
a not proper working of these elements.

While at it, the new layer required a fix for the format
in eglfs_kms.json. Otherwise, the ces-qt-browser wasn't able to
start.

Signed-off-by: Steffen Kothe <skothe@christ-es.de>